requirements.txt
Flask==3.0.2
requests==2.32.2
flask-socketio==5.3.4
torchaudio>=2.0.0
torch>=2.0.0
python-dotenv==1.0.1
pydub==0.25.1
